Added some fixes

+++ Diplomski_v1/Survey.Business/Models/ViewModels/DataView.cs
@@ -9,6 +9,6 @@ namespace Survey.Business.Models.ViewModels
     public class DataView
     {
         public string name { get; set; }
-        public int value { get; set; }
+        public decimal value { get; set; }
     }
 }
